# bpo-46939: Specialize calls to Python classes