--- a/Include/cpython/object.h
+++ b/Include/cpython/object.h
@@ -234,6 +234,7 @@ struct _typeobject {
  * by code other than the specializer and interpreter. */
 struct _specialization_cache {
     PyObject *getitem;
+    PyObject *init;
 };
 
 /* The *real* layout of a type object when allocated on the heap */