--- a/Include/internal/pycore_frame.h
+++ b/Include/internal/pycore_frame.h
@@ -61,6 +61,7 @@ typedef struct _PyInterpreterFrame {
     PyFrameState f_state;  /* What state the frame is in */
     bool is_entry;  // Whether this is the "root" frame for the current _PyCFrame.
     bool is_generator;
+    PyObject *self; /* Strong reference consumed by __init__ frames to return self in RETURN_VALUE */
     PyObject *localsplus[1];
 } _PyInterpreterFrame;
 
@@ -118,6 +119,7 @@ _PyFrame_InitializeSpecials(
     frame->f_state = FRAME_CREATED;
     frame->is_entry = false;
     frame->is_generator = false;
+    frame->self = NULL;
 }
 
 /* Gets the pointer to the locals array