--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -191,7 +191,7 @@ def_op('LIST_EXTEND', 162)
 def_op('SET_UPDATE', 163)
 def_op('DICT_MERGE', 164)
 def_op('DICT_UPDATE', 165)
-def_op('PRECALL', 166, 1)
+def_op('PRECALL', 166, 3)
 
 def_op('CALL', 171, 4)
 def_op('KW_NAMES', 172)