--- a/Python/specialize.c
+++ b/Python/specialize.c
@@ -591,15 +591,16 @@ initial_counter_value(void) {
 #define SPEC_FAIL_CALL_BAD_CALL_FLAGS 17
 #define SPEC_FAIL_CALL_CLASS 18
 #define SPEC_FAIL_CALL_PYTHON_CLASS 19
-#define SPEC_FAIL_CALL_METHOD_DESCRIPTOR 20
-#define SPEC_FAIL_CALL_BOUND_METHOD 21
-#define SPEC_FAIL_CALL_STR 22
-#define SPEC_FAIL_CALL_CLASS_NO_VECTORCALL 23
-#define SPEC_FAIL_CALL_CLASS_MUTABLE 24
-#define SPEC_FAIL_CALL_KWNAMES 25
-#define SPEC_FAIL_CALL_METHOD_WRAPPER 26
-#define SPEC_FAIL_CALL_OPERATOR_WRAPPER 27
-#define SPEC_FAIL_CALL_PYFUNCTION 28
+#define SPEC_FAIL_CALL_PYTHON_CLASS_NON_PY_INIT 20
+#define SPEC_FAIL_CALL_METHOD_DESCRIPTOR 21
+#define SPEC_FAIL_CALL_BOUND_METHOD 22
+#define SPEC_FAIL_CALL_STR 23
+#define SPEC_FAIL_CALL_CLASS_NO_VECTORCALL 24
+#define SPEC_FAIL_CALL_CLASS_MUTABLE 25
+#define SPEC_FAIL_CALL_KWNAMES 26
+#define SPEC_FAIL_CALL_METHOD_WRAPPER 27
+#define SPEC_FAIL_CALL_OPERATOR_WRAPPER 28
+#define SPEC_FAIL_CALL_PYFUNCTION 29
 
 /* COMPARE_OP */
 #define SPEC_FAIL_COMPARE_OP_DIFFERENT_TYPES 12
@@ -1523,7 +1524,33 @@ specialize_class_call(
     assert(_Py_OPCODE(*instr) == PRECALL_ADAPTIVE);
     PyTypeObject *tp = _PyType_CAST(callable);
     if (tp->tp_new == PyBaseObject_Type.tp_new) {
-        SPECIALIZATION_FAIL(PRECALL, SPEC_FAIL_CALL_PYTHON_CLASS);
+        _PyAdaptiveEntry *cache0 = &cache[0].adaptive;
+        _PyCallCache *cache1 = &cache[-1].call;
+        PyObject *descriptor = _PyType_Lookup(tp, &_Py_ID(__init__));
+        if (descriptor && Py_TYPE(descriptor) == &PyFunction_Type) {
+            if (!(tp->tp_flags & Py_TPFLAGS_HEAPTYPE)) {
+                return -1;
+            }
+            PyFunctionObject *func = (PyFunctionObject *)descriptor;
+            PyCodeObject *fcode = (PyCodeObject *)func->func_code;
+            int kind = function_kind(fcode);
+            if (kind != SIMPLE_FUNCTION) {
+                SPECIALIZATION_FAIL(PRECALL, kind);
+                return -1;
+            }
+            assert(tp->tp_version_tag != 0);
+            cache0->version = tp->tp_version_tag;
+            int version = _PyFunction_GetVersionForCurrentState(func);
+            if (version == 0 || version != (uint16_t)version) {
+                SPECIALIZATION_FAIL(PRECALL, SPEC_FAIL_OUT_OF_VERSIONS);
+                return -1;
+            }
+            cache1->func_version = version;
+            ((PyHeapTypeObject *)tp)->_spec_cache.init = descriptor;
+            *instr = _Py_MAKECODEUNIT(PRECALL_PY_CLASS, _Py_OPARG(*instr));
+            return 0;
+        }
+        SPECIALIZATION_FAIL(PRECALL, SPEC_FAIL_CALL_PYTHON_CLASS_NON_PY_INIT);
         return -1;
     }
     if (tp->tp_flags & Py_TPFLAGS_IMMUTABLETYPE) {