… Lib/test/test_dis.py
… Programs/test_frozenmain.h
… Python/opcode_targets.h