CALL_X supporting `__init__` inlining:
- [X] CALL_PY_EXACT_ARGS
- [X] CALL_PY_WITH_DEFAULTS
- [X] CALL

Micro benchmark using pyperf (Windows, PGO):
```
-m pyperf timeit -s "
>> class A:
>>  def __init__(self, a):
>>   self.a =a
>> " "A(1)" -o ..\pyperf-output\spec_py_class.json

Mean +- std dev: [spec_py_class_main] 159 ns +- 5 ns -> [spec_py_class] 106 ns +- 3 ns: 1.51x faster
```
<!-- issue-number: [bpo-46939](https://bugs.python.org/issue46939) -->
https://bugs.python.org/issue46939
<!-- /issue-number -->