@Fidget-Spinner 2022-03-06
I nearly went insane writing and debugging this. The general idea follows Mark's comments in https://github.com/faster-cpython/ideas/issues/267#issuecomment-1042942138. However, since we're specializing `__init__`, we need to return `self` instead of the `None` that `__init__` normally returns. This required a significant amount of hackery.