@Fidget-Spinner 2022-03-07
> This feels a bit fragile to me

Indeed, hence why I nearly went insane :). One very easy way to trip over is that all new `CALL_PY_*` opcodes must also handle this special case, like how the rest are currently doing so. I've added a `CALL_PY_FRAME_PASS_SELF()` macro so that it's more obvious to people. Most of our test suite will crash (I've tried this) if you forget to add it. Python class creation is an integral part of Python after all.