@Fidget-Spinner 2022-03-08
I've updated the first comment with a micro benchmark using pyperf. The results show `50%` speedup over `main`.